Add files via upload
Ayurvedic treatment to all age, Social, religion & Group. Currently, there is no primary Ayurveda Health clinic along with 
Dispensary in the nearby area. Over time, it has been observed that many patients are 
interested in getting Ayurvedic treatment for its rejuvenation values. 

With this background, the objective of the project is to set up an Ayurvedic Hospital at Garm 
Dongarpur, Gwalior. The location in front of Ohadpur reserve forest in the vicinity offers a serene natural setting.